feat: separate reth benchmark into a lib crate, generic over engine, and the binary (#393)
* Make `openvm-reth-benchmark` a lib crate

* Change .gitignore

* Make `HostArgs` a public struct

* Make `run_reth_benchmark` a public function

* `rustc` you understood me perfectly well

* Revert "Change .gitignore"

This reverts commit 5e1f4c6172595192f72432544610010224ba6209.

* Make client eth a parameter

* Switch to another openvm branch

* Fix stark-backend version and other minor things

* Bump openvm

* Return the binary back

* Update reth benchmark workflow

* Fix features propagation for the binary

* Maybe this time it's a fix idk

* chore: update openvm
