+++ bin/host/Cargo.toml
@@ -1,60 +1,25 @@
 [package]
-version = "0.1.0"
-name = "openvm-reth-benchmark"
-edition = "2021"
+name = "openvm-reth-benchmark-bin"
+edition.workspace = true
+repository.workspace = true
+homepage.workspace = true
+exclude.workspace = true
+version.workspace = true
 
 [dependencies]
-tokio.workspace = true
+clap_builder = "4.5.34"
 eyre.workspace = true
-url.workspace = true
-tracing-subscriber = "0.3.18"
-tracing.workspace = true
-dotenv = "0.15.0"
-clap = { version = "4.5.7", features = ["derive", "env"] }
-serde.workspace = true
-bincode = { workspace = true, features = ["std"] }
-metrics = "0.23.0"
-hex = "0.4.3"
-serde_json.workspace = true
+openvm-reth-benchmark = { workspace = true, default-features = false }
+openvm-stark-sdk.workspace = true
 
-# workspace
-openvm-host-executor.workspace = true
-openvm-client-executor.workspace = true
-
-# alloy
-alloy-provider.workspace = true
-alloy-rpc-client.workspace = true
-alloy-transport.workspace = true
-
-# reth
-reth-primitives.workspace = true
-
-# openvm
-openvm = { workspace = true }
-openvm-build = { workspace = true }
-openvm-transpiler = { workspace = true }
-openvm-rv32im-circuit = { workspace = true }
-openvm-algebra-circuit = { workspace = true }
-openvm-bigint-circuit = { workspace = true }
-openvm-ecc-circuit = { workspace = true }
-openvm-pairing-circuit = { workspace = true }
-openvm-circuit = { workspace = true }
-openvm-native-recursion = { workspace = true, features = ["static-verifier"] }
-openvm-benchmarks = { workspace = true, features = ["parallel"] }
-openvm-stark-sdk = { workspace = true }
-openvm-sdk = { workspace = true }
-openvm-native-compiler = { workspace = true }
-derive_more = "1.0.0"
-
-# halo2
-halo2-axiom = { version = "0.5.0", default-features = false }
+[lints]
+workspace = true
 
 [features]
-default = ["mimalloc", "bench-metrics"]
-bench-metrics = ["openvm-benchmarks/bench-metrics"]
-profiling = ["openvm-sdk/profiling", "openvm-benchmarks/profiling"]
-mimalloc = ["openvm-benchmarks/mimalloc"]
-jemalloc = ["openvm-benchmarks/jemalloc"]
-nightly-features = ["openvm-benchmarks/nightly-features"]
-# x86 only:
-halo2-asm = ["halo2-axiom/asm"]
+default = ["openvm-reth-benchmark/default"]
+bench-metrics = ["openvm-reth-benchmark/bench-metrics"]
+profiling = ["openvm-reth-benchmark/profiling"]
+mimalloc = ["openvm-reth-benchmark/mimalloc"]
+jemalloc = ["openvm-reth-benchmark/jemalloc"]
+nightly-features = ["openvm-reth-benchmark/nightly-features"]
+halo2-asm = ["openvm-reth-benchmark/halo2-asm"]
