+++ crates/executor/client/Cargo.toml
@@ -14,6 +14,7 @@ eyre.workspace = true
 serde_json.workspace = true
 serde.workspace = true
 tokio.workspace = true
+bincode_derive.workspace = true
 bincode.workspace = true
 rustc-hash.workspace = true
 hashbrown = { version = "0.14", features = ["rayon"] }
