+++ crates/host-bench/src/cli.rs
